Everything works!!
Cicl macros are defined using defcmacro.  The reason for this is to
allow reparsing after expansion, and to not conflict with Common Lisp.
The macro implementation isn't the greatest: macro names are
registered in the *cicl-macros* special variable.  Also, macros are
actually _functions_.  This has a few advantages, however...

The first is that the macros can affect the active lisp environment,
without generating code that must be "hidden" from the cicl parser.
allowing you to add arbitrary concepts which can be supported by lisp
data strutures.  For example, using macros it would be possible to add
some simple static polymorphism, or to implement type checking for
user defined functions.  Perhaps later I will add support for hooking
in more directly to keywords like `defun' used.